replace header.jsx file with header_simple_dark.jsx
